requirements.txt
USE FOLLLOWING COMMAND TO GENERATE THIS FILE:
pip3 install -r requirements.txt

OR USE THE FOLLOWING COMMAND TO GENERATE CERTAIN VERSIONS:
pip3 install package_name==version_number

OR USE THE FOLLOWING COMMAND TO INSTALL THE LATEST VERSION OF A CERTAIN PACKAGE:
pip3 install package_name

LIST OF REQUIRED PACKAGES AND THEIR VERSIONS:
python-docx==0.8.11
pyttsx3==2.90
requests==2.28.2
numpy==1.24.3
pandas==1.5.3
flask==2.2.3